+++ app/(routes)/packets/page.tsx
@@ -1,43 +1,47 @@
 'use client';
 
 import { useEffect, useState } from "react";
-import { 
-  Table,
-  flexRender,
-  Column,
-  ColumnDef,
+import {
   createColumnHelper,
-  useReactTable,
   getCoreRowModel,
   getFilteredRowModel,
-  getPaginationRowModel
-} from "@tanstack/react-table";
+  getPaginationRowModel,
+  useReactTable }
+from "@tanstack/react-table";
+import { PacketTable } from "./table";
+import { Popover } from "@headlessui/react";
 import { Packet, PacketStates } from "utils/types/packet";
-import { CHAIN_CONFIGS } from "utils/chains/configs";
+import { hideMiddleChars } from "utils/functions";
 
 const columnHelper = createColumnHelper<Packet>();
 const columns = [
   columnHelper.accessor('state', {
     header: 'State',
     cell: props => <span>{ stateToString(props.getValue()) }</span>,
-    enableColumnFilter: false
+    enableColumnFilter: false,
+    enableHiding: true
   }),
   columnHelper.accessor('sendTx', {
     header: 'Send Tx',
-    cell: props => hideMiddleChars(props.getValue())
+    cell: props => hideMiddleChars(props.getValue()),
+    enableHiding: true
   }),
   columnHelper.accessor('sourceChain', {
-    header: 'Source Chain'
+    header: 'Source Chain',
+    enableHiding: true
   }),
   columnHelper.accessor('destChain', {
-    header: 'Dest Chain'
+    header: 'Dest Chain',
+    enableHiding: true
   }),
   columnHelper.accessor('sourcePortAddress', {
     header: 'Src Port Address',
-    cell: props => hideMiddleChars(props.getValue())
+    cell: props => hideMiddleChars(props.getValue()),
+    enableHiding: true
   }),
   columnHelper.accessor('sourceChannelId', {
-    header: 'Src Channel ID'
+    header: 'Src Channel ID',
+    enableHiding: true
   }),
   // Create column that combines that subtracts difference between create and end time
   columnHelper.accessor(row => (row.endTime ?? 0) - row.createTime, {
@@ -48,14 +52,20 @@ const columns = [
       }
       return <span>{Math.round(props.getValue() / 1000)} s</span>;
     },
-    enableColumnFilter: false
-  })
-
+    enableColumnFilter: false,
+    enableHiding: true
+  }),
+  columnHelper.accessor('fee', {
+    header: 'Fee',
+    enableColumnFilter: false,
+    enableHiding: true
+  }),
 ];
 
 export default function Packets() {
   const [packets, setPackets] = useState<Packet[]>([]);
   const [loading, setLoading] = useState(true);
+  const [columnVisibility, setColumnVisibility] = useState({})
 
   useEffect(() => {
     fetch("/api/mock-packets").then(res => res.json()).then(data => {
@@ -72,26 +82,53 @@ export default function Packets() {
     });
   };
 
+  const table = useReactTable({
+    data: packets,
+    columns,
+    state: {
+      columnVisibility
+    },
+    onColumnVisibilityChange: setColumnVisibility,
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel()
+  })
+
   return (
     <div>
       <h2>Search for Packets:</h2>
+
+      <Popover className="relative">
+        <Popover.Button>Columns</Popover.Button>
+        <Popover.Panel className="absolute z-10">
+
+          <div className="bg-white dark:bg-bg-dk">
+            {table.getAllLeafColumns().map(column => { return (
+              <div key={column.id} className="px-1">
+                <label>
+                  <input
+                    {...{
+                      type: 'checkbox',
+                      checked: column.getIsVisible(),
+                      onChange: column.getToggleVisibilityHandler(),
+                    }}
+                  />{' '}
+                  {column.columnDef.header as string}
+                </label>
+              </div>
+              )
+            })}
+          </div>
+
+        </Popover.Panel>
+      </Popover>
+
       <button onClick={() => reloadData()} className="border p-2">
         Refresh
       </button>
-      {loading && <p>Loading...</p>}
-      {!loading && <>
-        <PacketTable {...{
-          data: packets,
-          columns
-        }} />
-      </> }
+      <PacketTable {...{table, loading}} />
     </div>
   );
-}
-
-function hideMiddleChars(str: string) {
-  const shortened =  str.slice(0, 7) + '...' + str.slice(-5);
-  return <span title={str}>{shortened}</span>;
 }
 
 function stateToString(state: PacketStates) {
@@ -100,145 +137,4 @@ function stateToString(state: PacketStates) {
     case PacketStates.POLY_WRITE_ACK || PacketStates.POLY_RECV: return 'Relaying';
     case PacketStates.RECV || PacketStates.SENT || PacketStates.WRITE_ACK: return 'Confirming';
   }
-}
-
-function PacketTable({ data, columns }: { data: Packet[], columns: ColumnDef<Packet, any>[] }) {
-  const table = useReactTable({
-    data,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getPaginationRowModel: getPaginationRowModel()
-  })
-
-  return (
-    <div>
-      <table>
-        <thead>
-          {table.getHeaderGroups().map(headerGroup => (
-            <tr key={headerGroup.id}>
-            {headerGroup.headers.map(header => {
-              return (
-                <th key={header.id} colSpan={header.colSpan} className="pl-8 first:pl-0">
-                  {header.isPlaceholder ? null : (
-                    <div>
-                      {flexRender(
-                        header.column.columnDef.header,
-                        header.getContext()
-                      )}
-                      {header.column.getCanFilter() ? (
-                        <div>
-                          <ColumnFilter column={header.column} table={table} />
-                        </div>
-                      ) : null}
-                    </div>
-                  )}
-                </th>
-              )
-            })}
-            </tr>
-          ))}
-        </thead>
-        <tbody>
-          {table.getRowModel().rows.map(row => (
-            <tr key={row.id}>
-              {row.getVisibleCells().map(cell => (
-                <td key={cell.id} className="pl-8 first:pl-0">
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <Pagination table={table} />
-    </div>
-  )
-}
-
-function ColumnFilter({ column, table }: { column: Column<any, any>, table: Table<any> }) {
-  const firstValue = table
-    .getPreFilteredRowModel()
-    .flatRows[0]?.getValue(column.id);
-
-  const columnFilterValue = column.getFilterValue();
-  
-  if (column.id === 'sourceChain' || column.id === 'destChain') {
-    return (
-      <select
-        value={(columnFilterValue ?? '') as string}
-        onChange={e => column.setFilterValue(e.target.value)}
-        className="w-36 border shadow rounded dark:bg-bg-dk"
-      >
-        <option value="">All</option>
-        {
-          Object.entries(CHAIN_CONFIGS).map(([key, value]) => (
-            <option key={key} value={key}>{ value.display }</option>
-          ))
-        }
-      </select>
-    );
-  } else if (typeof firstValue === 'string') {
-    return (
-      <input
-        type="text"
-        value={(columnFilterValue ?? '') as string}
-        onChange={e => column.setFilterValue(e.target.value)}
-        placeholder={`Search...`}
-        className="w-36 border shadow rounded dark:bg-bg-dk"
-      />
-    );
-  } else {
-    return null;
-  }
-}
-
-function Pagination({ table }: { table: Table<any> }) {
-  return (
-    <div className="flex items-center gap-2">
-      <button
-        className="border rounded p-1"
-        onClick={() => table.previousPage()}
-        disabled={!table.getCanPreviousPage()}>
-        {'<'}
-      </button>
-      <button
-        className="border rounded p-1"
-        onClick={() => table.nextPage()}
-        disabled={!table.getCanNextPage()}>
-        {'>'}
-      </button>
-      <span className="flex items-center gap-1">
-        <div>Page</div>
-        <strong>
-          {table.getState().pagination.pageIndex + 1} of {' '}
-          {table.getPageCount()}
-        </strong>
-      </span>
-      <span className="flex items-center gap-1">
-        | Go to page:
-        <input
-          type="number"
-          defaultValue={table.getState().pagination.pageIndex + 1}
-          onChange={e => {
-            const page = e.target.value ? Number(e.target.value) - 1 : 0
-            table.setPageIndex(page)
-          }}
-          className="border p-1 rounded w-16 dark:bg-bg-dk"
-        />
-      </span>
-      <select
-        value={table.getState().pagination.pageSize}
-        onChange={e => {
-          table.setPageSize(Number(e.target.value))
-        }}
-        className="dark:bg-bg-dk">
-        {[10, 20, 30, 40, 50].map(pageSize => (
-          <option key={pageSize} value={pageSize}>
-            Show {pageSize}
-          </option>
-        ))}
-      </select>
-    </div>
-  )
 }
