+++ app/utils/functions.tsx
@@ -5,3 +5,8 @@ export async function getLatestBlock(chainId: CHAIN) {
   const provider = new ethers.JsonRpcProvider(CHAIN_CONFIGS[chainId].rpc);
   return await provider.getBlock("latest");
 }
+
+export function hideMiddleChars(str: string) {
+  const shortened =  str.slice(0, 7) + '...' + str.slice(-5);
+  return <span title={str}>{shortened}</span>;
+}
