+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@cosmjs/stargate": "^0.32.2",
     "@cosmjs/tendermint-rpc": "^0.32.2",
+    "@headlessui/react": "^1.7.18",
     "@tanstack/react-table": "^8.12.0",
     "@types/flat-cache": "^2.0.2",
     "ethers": "^6.10.0",
